* Basic System::Call support when compiling with 64-bit MinGW/GCC toolchain * Win64 fixes
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6607 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/crc32.c
@@ -19,10 +19,11 @@
 #include "Platform.h"
 #include "crc32.h"
 #include "exehead/config.h"
+
 #ifdef NSIS_CONFIG_CRC_SUPPORT
 
 // this is based on the (slow,small) CRC32 implementation from zlib.
-crc32_t NSISCALL CRC32(crc32_t crc, const unsigned char *buf, unsigned int len)
+crc32_t NSISCALL CRC32(crc32_t crc, const unsigned char *buf, size_t len)
 {
     static crc32_t crc_table[256];
